requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# CLIP model
git+https://github.com/openai/CLIP.git

# Data processing and visualization
Pillow>=9.5.0
matplotlib>=3.7.0
imageio>=2.31.0

# Utilities
einops>=0.6.1
tqdm>=4.65.0

# Jupyter notebook support
jupyter>=1.0.0
ipython>=8.12.0
ipywidgets>=8.0.0

# Optional: For configuration management
pyyaml>=6.0
types-PyYAML>=6.0  # Type stubs for mypy

# Optional: For better performance
scikit-image>=0.21.0
